requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==3.0
pypdfium2==5.3.0
pyyaml==6.0.2
xmltodict==0.13.0